+++ src/main/java/com/backendoori/ootw/user/service/UserService.java
@@ -2,8 +2,8 @@ package com.backendoori.ootw.user.service;
 
 import java.util.Objects;
 import com.backendoori.ootw.common.AssertUtil;
-import com.backendoori.ootw.exception.AlreadyExistEmailException;
-import com.backendoori.ootw.exception.IncorrectPasswordException;
+import com.backendoori.ootw.user.exception.AlreadyExistEmailException;
+import com.backendoori.ootw.user.exception.IncorrectPasswordException;
 import com.backendoori.ootw.exception.UserNotFoundException;
 import com.backendoori.ootw.security.jwt.TokenProvider;
 import com.backendoori.ootw.user.domain.User;
@@ -14,7 +14,6 @@ import com.backendoori.ootw.user.dto.UserDto;
 import com.backendoori.ootw.user.repository.UserRepository;
 import com.backendoori.ootw.user.validation.Message;
 import com.backendoori.ootw.user.validation.PasswordValidator;
-import com.backendoori.ootw.user.validation.RFC5322;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
